Removed test from GHRepositoryTest and added to GHTagTest.
