+++ src/test/resources/org/kohsuke/github/GHTagTest/wiremock/testCreateTag/__files/orgs_github-api-test-org-d542cf64-ac9f-4fe7-bc38-9286b63c7828.json
@@ -1,0 +1,41 @@
+{
+  "login": "github-api-test-org",
+  "id": 7544739,
+  "node_id": "MDEyOk9yZ2FuaXphdGlvbjc1NDQ3Mzk=",
+  "url": "https://api.github.com/orgs/github-api-test-org",
+  "repos_url": "https://api.github.com/orgs/github-api-test-org/repos",
+  "events_url": "https://api.github.com/orgs/github-api-test-org/events",
+  "hooks_url": "https://api.github.com/orgs/github-api-test-org/hooks",
+  "issues_url": "https://api.github.com/orgs/github-api-test-org/issues",
+  "members_url": "https://api.github.com/orgs/github-api-test-org/members{/member}",
+  "public_members_url": "https://api.github.com/orgs/github-api-test-org/public_members{/member}",
+  "avatar_url": "https://avatars3.githubusercontent.com/u/7544739?v=4",
+  "description": null,
+  "is_verified": false,
+  "has_organization_projects": true,
+  "has_repository_projects": true,
+  "public_repos": 10,
+  "public_gists": 0,
+  "followers": 0,
+  "following": 0,
+  "html_url": "https://github.com/github-api-test-org",
+  "created_at": "2014-05-10T19:39:11Z",
+  "updated_at": "2015-04-20T00:42:30Z",
+  "type": "Organization",
+  "total_private_repos": 0,
+  "owned_private_repos": 0,
+  "private_gists": 0,
+  "disk_usage": 147,
+  "collaborators": 0,
+  "billing_email": "kk@kohsuke.org",
+  "default_repository_permission": "none",
+  "members_can_create_repositories": false,
+  "two_factor_requirement_enabled": false,
+  "plan": {
+    "name": "free",
+    "space": 976562499,
+    "private_repos": 0,
+    "filled_seats": 11,
+    "seats": 0
+  }
+}
