+++ src/test/resources/org/kohsuke/github/GHTagTest/wiremock/testCreateTag/__files/repos_github-api-test-org_github-api_git_refs-f654b3ef-74dc-4775-8809-f07c2a44925e.json
@@ -1,0 +1,10 @@
+{
+  "ref": "refs/tags/create_tag_test",
+  "node_id": "MDM6UmVmMjA2ODg4MjAxOmNyZWF0ZV90YWdfdGVzdA==",
+  "url": "https://api.github.com/repos/github-api-test-org/github-api/git/refs/tags/create_tag_test",
+  "object": {
+    "sha": "e7aa6d4afbaa48669f0bbe11ca3c4d787b2b153c",
+    "type": "tag",
+    "url": "https://api.github.com/repos/github-api-test-org/github-api/git/tags/e7aa6d4afbaa48669f0bbe11ca3c4d787b2b153c"
+  }
+}
