+++ src/test/resources/org/kohsuke/github/GHTagTest/wiremock/testCreateTag/__files/user-a8d3e54a-c069-4f34-9f86-e811649e0ce1.json
@@ -1,0 +1,33 @@
+{
+  "login": "martinvanzijl",
+  "id": 24422213,
+  "node_id": "MDQ6VXNlcjI0NDIyMjEz",
+  "avatar_url": "https://avatars0.githubusercontent.com/u/24422213?v=4",
+  "gravatar_id": "",
+  "url": "https://api.github.com/users/martinvanzijl",
+  "html_url": "https://github.com/martinvanzijl",
+  "followers_url": "https://api.github.com/users/martinvanzijl/followers",
+  "following_url": "https://api.github.com/users/martinvanzijl/following{/other_user}",
+  "gists_url": "https://api.github.com/users/martinvanzijl/gists{/gist_id}",
+  "starred_url": "https://api.github.com/users/martinvanzijl/starred{/owner}{/repo}",
+  "subscriptions_url": "https://api.github.com/users/martinvanzijl/subscriptions",
+  "organizations_url": "https://api.github.com/users/martinvanzijl/orgs",
+  "repos_url": "https://api.github.com/users/martinvanzijl/repos",
+  "events_url": "https://api.github.com/users/martinvanzijl/events{/privacy}",
+  "received_events_url": "https://api.github.com/users/martinvanzijl/received_events",
+  "type": "User",
+  "site_admin": false,
+  "name": "Martin van Zijl",
+  "company": null,
+  "blog": "",
+  "location": null,
+  "email": null,
+  "hireable": true,
+  "bio": "My strongest language is C++, but I also know C, Java, Python and SQL.",
+  "public_repos": 15,
+  "public_gists": 0,
+  "followers": 2,
+  "following": 0,
+  "created_at": "2016-12-07T00:18:58Z",
+  "updated_at": "2020-01-09T05:50:04Z"
+}
